refactor: simplify the OAuth2 demo with mock endpoints and update README
